"Realtime" simulation and reorganization
Eventhandler can now simulate the timing of events happening.

I relabeled all of the imports to have exact paths, instead of relative
paths because this is good practice.

I mixed node into networkobjects to reduce clutter. Networkobjects now
contains all the parents, and the specific implementations lie in host,
link, and router. This organization is good because host, links, and
routers are very different entities with very different logic.